Major changes

+++ Project/lib/matrix.h
@@ -1,7 +1,7 @@
 typedef struct Matrix{
   unsigned int col;
   unsigned int rows;
-  int **matrix; 
+  double **matrix; 
 }Matrix;
 
 int build_matrix( const unsigned int dimension_columns, const unsigned int dimension_rows, Matrix* mtx);
